+++ src/Twilio/Rest/Iam/V1/KeyResource.cs
@@ -29,7 +29,7 @@ using Twilio.Types;
 
 namespace Twilio.Rest.Iam.V1
 {
-    public class NewApiKeyResource : Resource
+    public class KeyResource : Resource
     {
     
 
@@ -47,7 +47,7 @@ namespace Twilio.Rest.Iam.V1
         }
 
         
-        private static Request BuildCreateRequest(CreateNewApiKeyOptions options, ITwilioRestClient client)
+        private static Request BuildCreateRequest(CreateKeyOptions options, ITwilioRestClient client)
         {
             
             string path = "/v1/Keys";
@@ -64,10 +64,10 @@ namespace Twilio.Rest.Iam.V1
         }
 
         /// <summary> Create a new Signing Key for the account making the request. </summary>
-        /// <param name="options"> Create NewApiKey parameters </param>
+        /// <param name="options"> Create Key parameters </param>
         /// <param name="client"> Client to make requests to Twilio </param>
-        /// <returns> A single instance of NewApiKey </returns>
-        public static NewApiKeyResource Create(CreateNewApiKeyOptions options, ITwilioRestClient client = null)
+        /// <returns> A single instance of Key </returns>
+        public static KeyResource Create(CreateKeyOptions options, ITwilioRestClient client = null)
         {
             client = client ?? TwilioClient.GetRestClient();
             var response = client.Request(BuildCreateRequest(options, client));
@@ -76,10 +76,10 @@ namespace Twilio.Rest.Iam.V1
 
         #if !NET35
         /// <summary> Create a new Signing Key for the account making the request. </summary>
-        /// <param name="options"> Create NewApiKey parameters </param>
+        /// <param name="options"> Create Key parameters </param>
         /// <param name="client"> Client to make requests to Twilio </param>
-        /// <returns> Task that resolves to A single instance of NewApiKey </returns>
-        public static async System.Threading.Tasks.Task<NewApiKeyResource> CreateAsync(CreateNewApiKeyOptions options, ITwilioRestClient client = null)
+        /// <returns> Task that resolves to A single instance of Key </returns>
+        public static async System.Threading.Tasks.Task<KeyResource> CreateAsync(CreateKeyOptions options, ITwilioRestClient client = null)
         {
             client = client ?? TwilioClient.GetRestClient();
             var response = await client.RequestAsync(BuildCreateRequest(options, client));
@@ -93,15 +93,15 @@ namespace Twilio.Rest.Iam.V1
         /// <param name="keyType">  </param>
         /// <param name="policy"> The \\\\`Policy\\\\` object is a collection that specifies the allowed Twilio permissions for the restricted key. For more information on the permissions available with restricted API keys, refer to the [Twilio documentation](https://www.twilio.com/docs/iam/api-keys/restricted-api-keys#permissions-available-with-restricted-api-keys). </param>
         /// <param name="client"> Client to make requests to Twilio </param>
-        /// <returns> A single instance of NewApiKey </returns>
-        public static NewApiKeyResource Create(
+        /// <returns> A single instance of Key </returns>
+        public static KeyResource Create(
                                           string accountSid,
                                           string friendlyName = null,
-                                          NewApiKeyResource.KeytypeEnum keyType = null,
+                                          KeyResource.KeytypeEnum keyType = null,
                                           object policy = null,
                                             ITwilioRestClient client = null)
         {
-            var options = new CreateNewApiKeyOptions(accountSid){  FriendlyName = friendlyName, KeyType = keyType, Policy = policy };
+            var options = new CreateKeyOptions(accountSid){  FriendlyName = friendlyName, KeyType = keyType, Policy = policy };
             return Create(options, client);
         }
 
@@ -112,29 +112,29 @@ namespace Twilio.Rest.Iam.V1
         /// <param name="keyType">  </param>
         /// <param name="policy"> The \\\\`Policy\\\\` object is a collection that specifies the allowed Twilio permissions for the restricted key. For more information on the permissions available with restricted API keys, refer to the [Twilio documentation](https://www.twilio.com/docs/iam/api-keys/restricted-api-keys#permissions-available-with-restricted-api-keys). </param>
         /// <param name="client"> Client to make requests to Twilio </param>
-        /// <returns> Task that resolves to A single instance of NewApiKey </returns>
-        public static async System.Threading.Tasks.Task<NewApiKeyResource> CreateAsync(
+        /// <returns> Task that resolves to A single instance of Key </returns>
+        public static async System.Threading.Tasks.Task<KeyResource> CreateAsync(
                                                                                   string accountSid,
                                                                                   string friendlyName = null,
-                                                                                  NewApiKeyResource.KeytypeEnum keyType = null,
+                                                                                  KeyResource.KeytypeEnum keyType = null,
                                                                                   object policy = null,
                                                                                     ITwilioRestClient client = null)
         {
-        var options = new CreateNewApiKeyOptions(accountSid){  FriendlyName = friendlyName, KeyType = keyType, Policy = policy };
+        var options = new CreateKeyOptions(accountSid){  FriendlyName = friendlyName, KeyType = keyType, Policy = policy };
             return await CreateAsync(options, client);
         }
         #endif
     
         /// <summary>
-        /// Converts a JSON string into a NewApiKeyResource object
+        /// Converts a JSON string into a KeyResource object
         /// </summary>
         /// <param name="json"> Raw JSON string </param>
-        /// <returns> NewApiKeyResource object represented by the provided JSON </returns>
-        public static NewApiKeyResource FromJson(string json)
+        /// <returns> KeyResource object represented by the provided JSON </returns>
+        public static KeyResource FromJson(string json)
         {
             try
             {
-                return JsonConvert.DeserializeObject<NewApiKeyResource>(json);
+                return JsonConvert.DeserializeObject<KeyResource>(json);
             }
             catch (JsonException e)
             {
@@ -185,7 +185,7 @@ namespace Twilio.Rest.Iam.V1
 
 
 
-        private NewApiKeyResource() {
+        private KeyResource() {
 
         }
     }