+++ src/Twilio/Rest/Verify/V2/Service/VerificationCheckResource.cs
@@ -101,6 +101,7 @@ namespace Twilio.Rest.Verify.V2.Service
         /// <param name="verificationSid"> A SID that uniquely identifies the Verification Check. Either this parameter or the `to` phone number/[email](https://www.twilio.com/docs/verify/email) must be specified. </param>
         /// <param name="amount"> The amount of the associated PSD2 compliant transaction. Requires the PSD2 Service flag enabled. </param>
         /// <param name="payee"> The payee of the associated PSD2 compliant transaction. Requires the PSD2 Service flag enabled. </param>
+        /// <param name="snaClientToken"> A sna client token received in sna url invocation response needs to be passed in Verification Check request and should match to get successful response. </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <returns> A single instance of VerificationCheck </returns>
         public static VerificationCheckResource Create(
@@ -110,9 +111,10 @@ namespace Twilio.Rest.Verify.V2.Service
                                           string verificationSid = null,
                                           string amount = null,
                                           string payee = null,
+                                          string snaClientToken = null,
                                             ITwilioRestClient client = null)
         {
-            var options = new CreateVerificationCheckOptions(pathServiceSid){  Code = code, To = to, VerificationSid = verificationSid, Amount = amount, Payee = payee };
+            var options = new CreateVerificationCheckOptions(pathServiceSid){  Code = code, To = to, VerificationSid = verificationSid, Amount = amount, Payee = payee, SnaClientToken = snaClientToken };
             return Create(options, client);
         }
 
@@ -124,6 +126,7 @@ namespace Twilio.Rest.Verify.V2.Service
         /// <param name="verificationSid"> A SID that uniquely identifies the Verification Check. Either this parameter or the `to` phone number/[email](https://www.twilio.com/docs/verify/email) must be specified. </param>
         /// <param name="amount"> The amount of the associated PSD2 compliant transaction. Requires the PSD2 Service flag enabled. </param>
         /// <param name="payee"> The payee of the associated PSD2 compliant transaction. Requires the PSD2 Service flag enabled. </param>
+        /// <param name="snaClientToken"> A sna client token received in sna url invocation response needs to be passed in Verification Check request and should match to get successful response. </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <returns> Task that resolves to A single instance of VerificationCheck </returns>
         public static async System.Threading.Tasks.Task<VerificationCheckResource> CreateAsync(
@@ -133,9 +136,10 @@ namespace Twilio.Rest.Verify.V2.Service
                                                                                   string verificationSid = null,
                                                                                   string amount = null,
                                                                                   string payee = null,
+                                                                                  string snaClientToken = null,
                                                                                     ITwilioRestClient client = null)
         {
-        var options = new CreateVerificationCheckOptions(pathServiceSid){  Code = code, To = to, VerificationSid = verificationSid, Amount = amount, Payee = payee };
+        var options = new CreateVerificationCheckOptions(pathServiceSid){  Code = code, To = to, VerificationSid = verificationSid, Amount = amount, Payee = payee, SnaClientToken = snaClientToken };
             return await CreateAsync(options, client);
         }
         #endif