+++ CHANGES.md
@@ -1,6 +1,22 @@
 twilio-csharp Changelog
 =======================
 
+[2024-10-03] Version 7.5.0
+--------------------------
+**Library - Feature**
+- [PR #760](https://github.com/twilio/twilio-csharp/pull/760): Merging cluster tests for orgs api. Thanks to [@AsabuHere](https://github.com/AsabuHere)!
+
+**Messaging**
+- Add A2P external campaign CnpMigration flag
+
+**Numbers**
+- Add address sid to portability API
+
+**Verify**
+- Add `SnaClientToken` optional parameter on Verification check.
+- Add `EnableSnaClientToken` optional parameter for Verification creation.
+
+
 [2024-09-25] Version 7.4.0
 --------------------------
 **Library - Feature**